tests: fixes
