+++ contracts/sdjwt-verifier/src/tests/mod.rs
@@ -1,4 +1,8 @@
 mod contract_test;
 mod fixtures;
 mod types;
+mod verifier_exp_test;
+mod verifier_register_test;
 mod verifier_test;
+mod verifier_update_test;
+mod verifier_verify_test;
